# Java/Lab6,7/InternetProvider/output/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:9000
  description: Generated server url
...
paths:
...
  /installation/update:
    put:
      tags:
      - Installation controller
      summary: Update client installation
      description: Changing client installation using new params
      operationId: updateInstallation
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: string
      - name: address
        in: query
        required: true
        schema:
          type: string
      - name: routerNumber
        in: query
        required: true
        schema:
          type: string
      - name: serviceId
        in: query
        required: true
        schema:
          type: string
      - name: clientId
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Installation updated
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Invalid params given
          content:
            '*/*':
              schema:
                type: string
  /client/update:
    put:
      tags:
      - Client controller
      summary: Update client
      description: Changing name and surname of client using id
      operationId: updateClient
      parameters:
      - name: number
        in: query
        required: true
        schema:
          type: string
      - name: name
        in: query
        required: true
        schema:
          type: string
      - name: surname
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Client updated
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Invalid params given
          content:
            '*/*':
              schema:
                type: string
...
  /installation/add:
    post:
      tags:
      - Installation controller
      summary: Add installation
      description: Adding installation to specified client using client id and installation
        params
      operationId: addInstallation
      parameters:
      - name: address
        in: query
        required: true
        schema:
          type: string
      - name: routerNumber
        in: query
        required: true
        schema:
          type: string
      - name: serviceId
        in: query
        required: true
        schema:
          type: string
      - name: clientId
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Installation added
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Invalid params given
          content:
            '*/*':
              schema:
                type: string
  /client/add:
    post:
      tags:
      - Client controller
      summary: Add client
      description: Adding client with specified name and surname
      operationId: addClient
      parameters:
      - name: name
        in: query
        required: true
        schema:
          type: string
      - name: surname
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Client added
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Invalid params given
          content:
            '*/*':
              schema:
                type: string
  /service/get/{id}:
    get:
      tags:
      - Service controller
      summary: Get service by id
      description: Get service using id
      operationId: getClientData
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Service fetched
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Invalid params given
          content:
            '*/*':
              schema:
                type: string
...
  /installation/get/{id}:
    get:
      tags:
      - Installation controller
      summary: Get client installations
      description: Fetching all client installations using client id
      operationId: getInstallationData
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Client installations fetched
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Invalid params given
          content:
            '*/*':
              schema:
                type: string
  /client/get/{id}:
    get:
      tags:
      - Client controller
      summary: Get client using id
      description: Fetching data of specified client using id
      operationId: getClientData_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Client fetched
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Invalid params given
          content:
            '*/*':
              schema:
                type: string
...
  /installation/delete:
    delete:
      tags:
      - Installation controller
      summary: Delete installation
      description: Delete installation using id
      operationId: deleteInstallation
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Installation deleted
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Invalid params given
          content:
            '*/*':
              schema:
                type: string
  /client/delete:
    delete:
      tags:
      - Client controller
      summary: Delete client
      description: Deleting client using id
      operationId: deleteClient
      parameters:
      - name: clientId
        in: query
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Invalid params given
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Client deleted
          content:
            '*/*':
              schema:
                type: string
...